requirements.txt
atlassian-python-api ~= 3.8.0
incqueryserver-api-python-client ~= 0.19.0.dev2
SPARQLWrapper ~= 1.8.5
lxml ~= 4.6.2
rdflib ~= 5.0.0
